Add empty environment variables for algolia
This adds empty environment variables to algolia if none are present.
When we are deploying to dev/preview environments on Vercel, the algolia
enviornment variables will be not defined and causes build errors. Here,
we just specify empty environment variables if they are not defined to
get around the build issues.